test(core): Isolate restore binary data ID in execution lifecycle hooks
