+++ packages/cli/test/unit/execution-lifecyle/restoreBinaryDataId.test.ts
@@ -1,0 +1,170 @@
+import { restoreBinaryDataId } from '@/executionLifecycleHooks/restoreBinaryDataId';
+import { BinaryDataService } from 'n8n-core';
+import { mockInstance } from '../../shared/mocking';
+
+import type { IRun } from 'n8n-workflow';
+import config from '@/config';
+
+function toIRun(item?: object) {
+	return {
+		data: {
+			resultData: {
+				runData: {
+					myNode: [
+						{
+							data: {
+								main: [[item]],
+							},
+						},
+					],
+				},
+			},
+		},
+	} as unknown as IRun;
+}
+
+function getDataId(run: IRun, kind: 'binary' | 'json') {
+	// @ts-ignore
+	return run.data.resultData.runData.myNode[0].data.main[0][0][kind].data.id;
+}
+
+const binaryDataService = mockInstance(BinaryDataService);
+
+for (const mode of ['filesystem-v2', 's3'] as const) {
+	describe(`on ${mode} mode`, () => {
+		beforeAll(() => {
+			config.set('binaryDataManager.mode', mode);
+		});
+
+		afterEach(() => {
+			jest.clearAllMocks();
+		});
+
+		it('should restore if binary data ID is missing execution ID', async () => {
+			const workflowId = '6HYhhKmJch2cYxGj';
+			const executionId = 'temp';
+			const binaryDataFileUuid = 'a5c3f1ed-9d59-4155-bc68-9a370b3c51f6';
+
+			const incorrectFileId = `workflows/${workflowId}/executions/temp/binary_data/${binaryDataFileUuid}`;
+
+			const run = toIRun({
+				binary: {
+					data: { id: `s3:${incorrectFileId}` },
+				},
+			});
+
+			await restoreBinaryDataId(run, executionId, 'webhook');
+
+			const correctFileId = incorrectFileId.replace('temp', executionId);
+			const correctBinaryDataId = `s3:${correctFileId}`;
+
+			expect(binaryDataService.rename).toHaveBeenCalledWith(incorrectFileId, correctFileId);
+			expect(getDataId(run, 'binary')).toBe(correctBinaryDataId);
+		});
+
+		it('should do nothing if binary data ID is not missing execution ID', async () => {
+			const workflowId = '6HYhhKmJch2cYxGj';
+			const executionId = '999';
+			const binaryDataFileUuid = 'a5c3f1ed-9d59-4155-bc68-9a370b3c51f6';
+
+			const fileId = `workflows/${workflowId}/executions/${executionId}/binary_data/${binaryDataFileUuid}`;
+
+			const binaryDataId = `s3:${fileId}`;
+
+			const run = toIRun({
+				binary: {
+					data: {
+						id: binaryDataId,
+					},
+				},
+			});
+
+			await restoreBinaryDataId(run, executionId, 'webhook');
+
+			expect(binaryDataService.rename).not.toHaveBeenCalled();
+			expect(getDataId(run, 'binary')).toBe(binaryDataId);
+		});
+
+		it('should do nothing if no binary data ID', async () => {
+			const executionId = '999';
+			const dataId = '123';
+
+			const run = toIRun({
+				json: {
+					data: { id: dataId },
+				},
+			});
+
+			await restoreBinaryDataId(run, executionId, 'webhook');
+
+			expect(binaryDataService.rename).not.toHaveBeenCalled();
+			expect(getDataId(run, 'json')).toBe(dataId);
+		});
+
+		it('should do nothing on itemless case', async () => {
+			const executionId = '999';
+
+			const promise = restoreBinaryDataId(toIRun(), executionId, 'webhook');
+
+			await expect(promise).resolves.not.toThrow();
+
+			expect(binaryDataService.rename).not.toHaveBeenCalled();
+		});
+
+		it('should do nothing if data is undefined', async () => {
+			const executionId = '999';
+
+			const run = toIRun({
+				json: {
+					data: undefined,
+				},
+			});
+
+			const promise = restoreBinaryDataId(run, executionId, 'webhook');
+
+			await expect(promise).resolves.not.toThrow();
+
+			expect(binaryDataService.rename).not.toHaveBeenCalled();
+		});
+
+		it('should do nothing if workflow execution mode is not `webhook`', async () => {
+			const executionId = '999';
+
+			const run = toIRun({
+				json: {
+					data: undefined,
+				},
+			});
+
+			const promise = restoreBinaryDataId(run, executionId, 'internal');
+
+			await expect(promise).resolves.not.toThrow();
+
+			expect(binaryDataService.rename).not.toHaveBeenCalled();
+		});
+	});
+}
+
+describe('on default mode', () => {
+	afterEach(() => {
+		config.load(config.default);
+	});
+
+	it('should do nothing', async () => {
+		config.set('binaryDataManager.mode', 'default');
+
+		const executionId = '999';
+
+		const run = toIRun({
+			json: {
+				data: undefined,
+			},
+		});
+
+		const promise = restoreBinaryDataId(run, executionId, 'internal');
+
+		await expect(promise).resolves.not.toThrow();
+
+		expect(binaryDataService.rename).not.toHaveBeenCalled();
+	});
+});
